pyproject.toml
[project]
name = "pytuin_desktop"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
]

[project.scripts]
demo   = "examples.demo:main"
demo-edit = "examples.demo_editor:main"
config = "config:print_envs"
main   = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/examples"]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
